backend/pyproject.toml
[project]
name = "config"
version = "0.1.0"
description = ""
authors = [
    {name = "Praveenr-2101",email = "prabhapraveen000@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
